+++ src/lib/blendings/BlendingAlgorithms.cpp
@@ -80,7 +80,7 @@ qreal LinearLightBlending::blendChannel( qreal const bottomColorIntensity,
                                          qreal const topColorIntensity ) const
 {
     return qMin( qreal( 1.0 ),
-                 qMax( qreal( 0.0 ), ( bottomColorIntensity + 2.0 * topColorIntensity ) - 1.0 ));
+                 qMax( qreal( 0.0 ), qreal( bottomColorIntensity + 2.0 * topColorIntensity - 1.0 )));
 }
 
 qreal OverlayBlending::blendChannel( qreal const bottomColorIntensity,
@@ -120,7 +120,7 @@ qreal ColorBurnBlending::blendChannel( qreal const bottomColorIntensity,
     Q_UNUSED(topColorIntensity);
     // FIXME: check if this formula makes sense
     return qMin( qreal( 1.0 ),
-                 qMax( qreal( 0.0 ), 1.0 - ( 1.0 - bottomColorIntensity ) / topColorIntensity ));
+                 qMax( qreal( 0.0 ), qreal( 1.0 - ( 1.0 - bottomColorIntensity ) / topColorIntensity )));
 }
 
 qreal DarkBlending::blendChannel( qreal const bottomColorIntensity,
@@ -201,7 +201,7 @@ qreal ColorDodgeBlending::blendChannel( qreal const bottomColorIntensity,
                                         qreal const topColorIntensity ) const
 {
     return qMin( qreal( 1.0 ),
-                 qMax( qreal( 0.0 ), bottomColorIntensity / ( 1.0 - topColorIntensity )));
+                 qMax( qreal( 0.0 ), qreal( bottomColorIntensity / ( 1.0 - topColorIntensity ))));
 }
 
 qreal GammaLightBlending::blendChannel( qreal const bottomColorIntensity,
@@ -255,9 +255,9 @@ qreal VividLightBlending::blendChannel( qreal const bottomColorIntensity,
 {
     return topColorIntensity < 0.5
         ? qMin( qreal( 1.0 ),
-                qMax( qreal( 0.0 ), 1.0 - ( 1.0 - bottomColorIntensity ) / ( 2.0 * topColorIntensity )))
+                qMax( qreal( 0.0 ), qreal( 1.0 - ( 1.0 - bottomColorIntensity ) / ( 2.0 * topColorIntensity ))))
         : qMin( qreal( 1.0 ),
-                qMax( qreal( 0.0 ), bottomColorIntensity / ( 2.0 * ( 1.0 - topColorIntensity ))));
+                qMax( qreal( 0.0 ), qreal( bottomColorIntensity / ( 2.0 * ( 1.0 - topColorIntensity )))));
 }
 
 
@@ -284,7 +284,7 @@ qreal BleachBlending::blendChannel( qreal const bottomColorIntensity,
 qreal DifferenceBlending::blendChannel( qreal const bottomColorIntensity,
                                         qreal const topColorIntensity ) const
 {
-    return qMax( qMin( qreal( 1.0 ), bottomColorIntensity - topColorIntensity + 0.5 ),
+    return qMax( qMin( qreal( 1.0 ), qreal( bottomColorIntensity - topColorIntensity + 0.5 )),
                  qreal( 0.0 ));
 }
